update certificate component
